requirements.txt
pyyaml==6.0.2
colorzero==2.0
seaborn==0.13.2
drawsvg==2.4.1
